% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connections.R
\name{Autowire}
\alias{Autowire}
\title{Autowire Directed Unpack Connections}
\usage{
Autowire(headModules, tailModules, connectionNamePrefix = NULL,
  exclusions = c(""))
}
\arguments{
\item{headModules}{a module object, list of module objects, or character vector of names of modules.
If multiple module objects are provided and several of them output variables with the same name
Output of earlier modules in the list will be ignored when constructing a connection.}

\item{tailModules}{a module object, list of module objects, or character vector of names of modules}

\item{connectionNamePrefix}{a character string to use as a prefix in the
name generation of connections created with 
this function. Default is \code{NULL}, in which case
The head and tail module names are concatenated
to be used as the prefix.}

\item{exclusions}{a character vector of named outputs from
\code{headModules} to exclude from autowiring.}
}
\value{
list of created connections
}
\description{
Creates directed unpack connections automatically between
             \code{headModules} and \code{tailModules}. For this to work,
             the associated function object of a head module must have a
             single extractable return statement that returns a named list.
             When this returned named list has a name collision with an input
             argument name of a tail module, a
             \code{DirectedUnpackConnection} is created between the
             head module and the tail Module. An error is thrown by this
             function when the associated function obj of a head module
             is not set up for autowiring. This specifically happens when
             the associated function obj has 0 or multiple extractable
             return statements in the body of the associated function
             object, or the return statement does not return a named list.
             Note, the absence of an error message does not imply
             connections were created -- just that the \code{headModules}
             return statements are compatible with this autowiring function.
}
