% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_interface.R
\name{CheckInterfaceImplementation}
\alias{CheckInterfaceImplementation}
\title{CheckInterfaceImplementation}
\usage{
CheckInterfaceImplementation(R6Class)
}
\arguments{
\item{R6Class}{an R6 class in this library or a character string of the class
name in this library.}
}
\value{
NULL if successful. Raises a fatal error if finds a missing method
        expected in \code{R6Class}.
}
\description{
Checks to see if \code{R6Class} has fully implemented
             an interface contract.
}
