% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_execution.R
\name{CombineArgLists}
\alias{CombineArgLists}
\title{Combine Argument Lists}
\usage{
CombineArgLists(args1, args2)
}
\arguments{
\item{args1}{named list of function arguments. Has higher precedence over
\code{args2} in the combined list.}

\item{args2}{named list of function arguments.}
}
\value{
list
}
\description{
combines arguments lists \code{args1} and \code{arg2} into one
             list. If there is a name collision between \code{args1} and
             \code{arg2}, the \code{arg1} value will be in the returned
             list.
}
