% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_interface.R
\docType{data}
\name{ConnectionInterface}
\alias{ConnectionInterface}
\title{Connection Interface}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConnectionInterface
}
\description{
Defines a contract of implementation for any connection based class, i.e. a class that extends this
             interface. This class cannot be instantiated directly as it is abstract.
}
\section{Contracted Methods}{

\describe{
   \item{\code{filterOutputValue(outputValue)}}{filters the \code{outputValue}, value outputted from the source module of a conntion, to what is expected by the destination module.}
   \item{\code{getInputArgument()}}{Gets the argument the output of one module should be used as input into the other module's execution.}
}
}

\keyword{datasets}
