% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connections.R
\name{CreateDirectedConnections}
\alias{CreateDirectedConnections}
\title{Create Directed Connections}
\usage{
CreateDirectedConnections(headModules, tailModules, inputArgument,
  connectionNamePrefix = NULL, connectionType = DirectedConnection)
}
\arguments{
\item{headModules}{a list of module objs or names of modules}

\item{tailModules}{a list of modules objs or names of modules}

\item{inputArgument}{Name of the argument
the output of a head module is used for a tail
module's execution.}

\item{connectionNamePrefix}{a character string to use as a prefix in the
name generation of connections created with 
this function. Default is \code{NULL}, in which case
The head and tail module names are concatenated
to be used as the prefix.}

\item{connectionType}{R6Class that is of \code{DirectedConnection} type,
any classes that inherit from
\code{DirectedConnection} will work.}
}
\value{
list of directed connections created
}
\description{
creates a list of directed connections of \code{connectionType}
             type from all \code{headModules} to all \code{tailModules}.
}
