% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_custom_function.R
\name{CustomFunctionModule$initFromSaveData}
\alias{CustomFunctionModule$initFromSaveData}
\title{CustomFunctionModule Init From Save Data}
\arguments{
\item{saveData}{json saved data generated through the method \code{getSaveInfo}.}
}
\description{
Static factory method that generates a CustomFunctionModule
             from json save data.
}
