% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dag.R
\name{DAGWorkflow$initFromFile}
\alias{DAGWorkflow$initFromFile}
\title{DAG Workflow initFromFile Static Class Method}
\arguments{
\item{filename}{file to initialize from}
}
\value{
DAGWorkflow obj
}
\description{
static factory method to initialize a \code{DAGWorkflow} object from a file
}
