% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_directed.R
\name{DirectedConnection$initFromSaveData}
\alias{DirectedConnection$initFromSaveData}
\title{DirectedConnection Init From Save Data}
\arguments{
\item{saveData}{json saved data generated through the method \code{getSaveInfo}.}
}
\description{
Static factory method that generates a DirectedConnection
             from json save data.
}
