% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_directed_unpack.R
\name{DirectedUnpackConnection$initFromSaveData}
\alias{DirectedUnpackConnection$initFromSaveData}
\title{DirectedUnpackConnection Init From Save Data}
\arguments{
\item{saveData}{json saved data generated through the method \code{getSaveInfo}.}
}
\description{
Static factory method that generates a DirectedUnpackConnection
             from json save data.
}
