% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_directed_unpack.R
\docType{data}
\name{DirectedUnpackConnection}
\alias{DirectedUnpackConnection}
\title{Directed Unpack Connection}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
DirectedUnpackConnection
}
\description{
Implementation of \code{ConnectionInterface} which represents a
             directed flow of info between a head module to a tail module.
             This is an extension of \code{DirectedConnection} designed to
             unpack a specific named value out of a list outputted by the
             head module as input into the tail module's execution rather
             than using the direct output of the head module as input into
             the tail module.
}
\section{Class Constructor}{

\describe{
    \item{\code{new(name, headModule, tailModule, inputArgument)}}{
        \itemize{
            \item{\bold{\code{name}}: unique character string that identifies
                   this connection.}
            \item{\bold{\code{headModule}}: obj that implements
                   \code{ModuleInterface} or a character string that
                    represents the head module of this connection.}
            \item{\bold{\code{tailModule}}: obj that implements
                   \code{ModuleInterface} or a character string that
                    represents thetail module of this connection.}
            \item{\bold{\code{inputArgument}}: name of the inputArgument
                  the output of the head module is used for in the tail
                  module's execution. Should be named with what value should
                  be unpacked out of the head module's return list and used
                  for the tail module's execution.}
        }
    }
}
}

\section{Inheritance From}{

\describe{
    \item{\link[updraft]{DirectedConnection}}{}
}
}

\section{Public}{

\describe{
    \item{\code{filterOutputValue(outputValue)}}{
        \itemize{
            \item{Filters the output of the \code{headModule}, \code{outputValue}, to what is expected by the \code{tailModule}. With this implementation of a directed connection, assumes \code{outputValue} is a list and the associated name of \code{inputArgument} is used to filter down \code{outputValue}.}
            \item{\bold{Returns}: filtered \code{outputValue}}
        }
    }
}
}

\section{Static Class Methods}{

\describe{
    \item{\link[updraft]{DirectedUnpackConnection-cash-initFromSaveData}}{}
}
}

\keyword{datasets}
