% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{Execute}
\alias{Execute}
\title{Execute a Workflow}
\usage{
Execute(workflow, moduleArgs = NULL, argsContainer = NULL,
  mode = SERIAL_MODE, clearCache = TRUE, moduleInputsSaveDir = NULL)
}
\arguments{
\item{workflow}{implementation instance of \code{WorkflowInterface}}

\item{moduleArgs}{a list of lists where the name of each element corresponds
to a module name in \code{workflow} and the elements are
argument lists to be used for arguments not fulfilled by
workflow connections to execute a module. Note, values
passed with an actual connection in the workflow will
override anything specified in this list.}

\item{argsContainer}{a list or environment of named values to use whenever
the name of a value in the container interects with
an input argument of a module in \code{workflow}. Note,
values passed with an actual connection in
\code{workflow} or specified by \code{moduleArgs}
override anything specified by this container.}

\item{mode}{Defines execution mode, e.x. running modules in parallel. All
defined modes are in the constants of this package.}

\item{clearCache}{boolean flag that indicated if will clear out workflow
module output caches.}

\item{moduleInputsSaveDir}{Name of directory to save module inputs to.
Inputs are saved in .rds format in form of
'module_name' + '_inputs.rds'. If \code{NULL}
, the default, inputs are not saved. Please note,
this feature is intended for debugging purposes
-- will slow execution down and take up disk
space.}
}
\description{
executes a workflow
}
