% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connections.R
\name{ExtractReturnStatements}
\alias{ExtractReturnStatements}
\title{Extract Return Statements of a Function}
\usage{
ExtractReturnStatements(func)
}
\arguments{
\item{func}{R function object}
}
\value{
a list of return calls
}
\description{
returns all return calls associated with \code{func}
}
