% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modules.R
\name{FutureFunctionCall}
\alias{FutureFunctionCall}
\title{Future Function Call}
\usage{
FutureFunctionCall(func, args, funcName = NULL, assignedProcesses = 1)
}
\arguments{
\item{func}{a R function object}

\item{args}{arguments to be used when calling \code{func}}

\item{funcName}{the name of the function being passed to this call, in case an error occurs it will log out the stack trace separately}

\item{assignedProcesses}{number of processes to assign to the execution of
this module. If the underlying execution code
leverages the package \code{future}, this will be
enforced by setting a future plan on behalf of the
module execution code.}
}
\value{
created future.
}
\description{
Creates a future for do.call of \code{func} with \code{args}.
             Wraps do.call so that, if an error occurs, the stack trace
             and error message get logged appriopriately.
}
