% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_interface.R
\name{is.Connection}
\alias{is.Connection}
\title{is.Connection}
\usage{
is.Connection(connection)
}
\arguments{
\item{connection}{possible connection obj}
}
\value{
boolean where \code{TRUE} indicates \code{connection} is a valid 
        connection object
}
\description{
Returns a boolean that indicates if \code{connection} is a 
             valid connection obj
}
