% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_directed.R
\name{is.DirectedConnection}
\alias{is.DirectedConnection}
\title{is Directed Connection}
\usage{
is.DirectedConnection(obj)
}
\arguments{
\item{obj}{object instance}
}
\value{
boolean where \code{TRUE} indicates \code{obj} is a \code{DirectedConnection}
        and \code{FALSE} otherwise.
}
\description{
Returns a boolean indicating if \code{obj} is a valid
             \code{DirectedConnection}.
}
