% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_interface.R
\name{is.Module}
\alias{is.Module}
\title{is.Module}
\usage{
is.Module(module)
}
\arguments{
\item{module}{An object to check}
}
\value{
boolean where \code{TRUE} indicates \code{module} is a valid 
        module object
}
\description{
Returns a boolean that indicates if \code{module} is a 
             valid module object. A valid module object is one that implements
             the interface defined by \code{\link{ModuleInterface}}.
}
