% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_interface.R
\name{is.Workflow}
\alias{is.Workflow}
\title{is Workflow}
\usage{
is.Workflow(obj)
}
\arguments{
\item{obj}{object instance}
}
\value{
boolean where \code{TRUE} indicates \code{obj} is a valid
        implementation instance of \code{WorkflowInterface}
        and \code{FALSE} otherwise.
}
\description{
Returns a boolean indicating if \code{obj} is a valid
             implementation instance of \code{Workflow}.
}
