% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_interface.R
\docType{data}
\name{ModuleInterface}
\alias{ModuleInterface}
\title{Module Interface}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModuleInterface
}
\description{
Defines a contract of implementation for any module based class, i.e. a class that extends this
             interface. This class cannot be instantiated directly as it is abstract.
}
\section{Contracted Methods}{

\describe{
     \item{\code{clearOutputCache()}}{Clears out the cached output of the associated module.}
     \item{\code{getFuncObj()}}{Gets the base R function object that is called when the associated module's execution is started.}
     \item{\code{getInput()}}{Gets a named logical vector where the names are the possible inputs into a module and the values indicate if an input is required.}
     \item{\code{getOutput()}}{Gets the output of the associated module.}
     \item{\code{hasCompleted()}}{Indicates if the associated module's execution has completed.}
     \item{\code{startExecution(argsList)}}{Starts the execution of the associated module with \code{argsList} as arguments.}
}
}

\seealso{
is.Module
}
\keyword{datasets}
