% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_package_function.R
\name{PackageFunctionModule$initFromSaveData}
\alias{PackageFunctionModule$initFromSaveData}
\title{PackageFunctionModule Init From Save Data}
\arguments{
\item{saveData}{json saved data generated through the method \code{getSaveInfo}.}
}
\value{
initialized \code{PackageFunctionModule}
}
\description{
Static factory method that generates a PackageFunctionModule
             from json save data.
}
