% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PARALLEL_MODE}
\alias{PARALLEL_MODE}
\title{Parallel Mode Constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
PARALLEL_MODE
}
\description{
value to set \code{mode} in \code{Execution} to run in parallel mode
}
\keyword{datasets}
