% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SERIAL_MODE}
\alias{SERIAL_MODE}
\title{Serial Mode Constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
SERIAL_MODE
}
\description{
value to set \code{mode} in \code{Execution} to run in serial mode
}
\keyword{datasets}
