% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_execution.R
\name{SetExecutionMode}
\alias{SetExecutionMode}
\alias{SetExecutionSettings}
\title{Set Execution Mode}
\usage{
SetExecutionSettings(mode, moduleMaxMemorySize)
}
\arguments{
\item{mode}{Execution mode. Must use a mode that is defined in UpDraft's
constants.}

\item{moduleMaxMemorySize}{Max memory a module can occupy without causing
a runtime error. Value is assumed to be in
Bytes.}
}
\value{
\code{NULL}
}
\description{
Sets the execution settings of a workffow. Note, this is done
             by ajusting settings of the \code{future} package. This might
             affect calling programs. See package documentation/vignettes
             for info.
}
