% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{UpDraftSettings}
\alias{UpDraftSettings}
\title{UpDraft Settings}
\format{An object of class \code{UpDraftSettings} (inherits from \code{R6}) of length 13.}
\usage{
UpDraftSettings
}
\description{
Stores UpDraft settings. User can directly adjust settings in
             this object to inject things into UpDraft such as custom logger
             functions.
}
\section{Adjustable Settings}{

\describe{
    \item{\code{errorLogger}}{
        \itemize{
            \item{function to use for UpDraft error statements. Specified function should allow \code{...} string inputs. Caution, specified function should throw a standard R error when called.}
         }
    }
    \item{\code{infoLogger}}{
        \itemize{
            \item{function to use for UpDraft info statements. Specified function should allow \code{...} string inputs.}
        }
    }
    \item{\code{moduleMaxMemorySize}}{
        \itemize{
            \item{Max memory a module can occupy without causing a runtime error. Value is assumed to be in Bytes.}
        }
    }
    \item{\code{moduleBorderColor}}{
        \itemize{
            \item{Graphviz color string of color to render the border of modules in UpDraft workflow visualization.}
        }
    }
    \item{\code{moduleFillColor}}{
        \itemize{
            \item{Graphviz color string of color to render the fill of modules in UpDraft workflow visualization.}
        }
    }
    \item{\code{monitorRate}}{
        \itemize{
            \item{Numeric value in seconds for the amount of time the monitoring process sleeps between checking if modules have completed.}
        }
    }
    \item{\code{multipleConnectionsColor}}{
        \itemize{
            \item{Graphviz color string of color to render multiple connections in UpDraft workflow visualization.}
        }
    }
    \item{\code{singleConnectionColor}}{
        \itemize{
            \item{Graphviz color string of color to render single connections in UpDraft workflow visualization.}
        }
    }
    \item{\code{warnLogger}}{
        \itemize{
            \item{function to use for UpDraft warning statements. Specified function should allow \code{...} string inputs. Caution, specified function should throw a standard R warning when called.}
        }
    }
}
}

\section{Public}{

\describe{
    \item{\code{update(...)}}{
        \itemize{
            \item{Updates multiple settings at once.}
            \item{\bold{\code{...}}: named parameters where the name of each parameter is the setting to change and the value of the parameter is the setting's new value.}
            \item{\bold{Returns}: \code{NULL}}
        }
    }
}
}

\section{Private}{

\describe{
    \item{\code{values}}{
        \itemize{
            \item{Stores values of settings.}
        }
    }
    \item{\code{setFunction(value, setting)}}{
        \itemize{
            \item{Updates a setting that stores a function type obj. This is a helper function for this object's active bindings.}
            \item{\bold{\code{value}}: value to update setting with. If \code{NULL}, signifies to return current value of \code{setting}.}
            \item{\bold{\code{setting}}: string of setting to update.}
            \item{\bold{Returns}: If \code{value} is \code{NULL}, returns current setting value. Else, updates setting to \code{value}.}
        }
    }
    \item{\code{setPositiveNumeric(value, setting)}}{
        \itemize{
            \item{Updates a setting that stores a positive numeric. This is a helper function for this object's active bindings.}
            \item{\bold{\code{value}}: value to update setting with. If \code{NULL}, signifies to return current value of \code{setting}.}
            \item{\bold{\code{setting}}: string of setting to update.}
            \item{\bold{Returns}: If \code{value} is \code{NULL}, returns current setting value. Else, updates setting to \code{value}.}
        }
    }
    \item{\code{setString(value, setting)}}{
        \itemize{
            \item{Updates a setting that stores a single string. This is a helper function for this object's active bindings.}
            \item{\bold{\code{value}}: value to update setting with. If \code{NULL}, signifies to return current value of \code{setting}.}
            \item{\bold{\code{setting}}: string of setting to update.}
            \item{\bold{Returns}: If \code{value} is \code{NULL}, returns current setting value. Else, updates setting to \code{value}.}
        }
    }
}
}

\keyword{datasets}
