% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_interface.R
\name{WorkflowInterface$initFromFile}
\alias{WorkflowInterface$initFromFile}
\title{WorkflowInterface Init From File}
\arguments{
\item{filename}{json file of saved workflow from \code{save} method.}
}
\value{
initialized \code{PackageFunctionModule}
}
\description{
Static factory method that generates a workflow object from
             json file.
}
